backend/django/requirements.txt
django==5.0.6
djangorestframework==3.15.2
Pillow==10.3.0
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
opencv-python-headless==4.8.1.78
psutil==5.9.6

# AI Model Dependencies
torch==2.4.1
torchvision==0.19.1
transformers==4.35.0
ultralytics>=8.0.200
numpy>=1.26.0
scipy==1.11.3
scikit-image==0.21.0
rembg>=2.0.55
onnxruntime>=1.16.0
whitenoise==6.6.0
